Rouster activity added to the stock Moonlight
This app has the ability to connect to both:
 - NVidia GeForce Experience GameStream
 - Sunshine

No matter of which one of them is chosen, a separate app to auto-
pair should be developed, which is not included here.

In order to create a Rouster app we actually fork the Moonligh app
and add the Rouster interface to it. To keep Rouster app as easy as
possible to be rebased with Moonlight we are trying to keep the
changes at minimum and extract as much as possible as a separate
files.

Here apart from slight changes in the build.gradle and manifest we
mainly introduce a simple activity to take care of the whole Rouster
process.

The activity basically introduce a webview showing a React based app
(a.k.a. the Rouster hybrid app).
 - Authorization and registration
 - VM administration
 - Safe mode

When the Hybrid need to pass operations on the Android app, it
hits certain URLs with the according parameters. The webView on the
Android app is monitoring certain URLs and when some of them is hit
it routes to an Android based routines and possibly opening Moonlight
activities.

We curretly make active use of two of the moonlight activities:
 - Game - to start streaming
 - Settings - to be able to set inside Moonligh all the streaming
   aspects

Apart from this we actively use some of the Moonlight services in
order to be able to automatically:
 - Add computer based on the curretly logged user
 - Pair Moonlight with the Android device
 - Receive information from the host, about it
 - Start streaming

All the implementation is curretly in a single activity (RousterMain)
file with just a simple router-like structure and methods for the
various tasks it need to do in order to pass the streaming to the
Moonlight Game activity. After we pass to activity we unbind the
service and we bind again only if we need to restart the stream.