Initial Amplify integration with Cognito
For the cognito implementation we use the Amplify provided instrumentarium, as
it provdes a bit easier integration "out of the box". Which is only true if we
discard the whole "poor documentation" and "code generation" bloodbath that is
accompanying the whole process.

There is a short code fragment in RosterMain.java for logging out the user in
case we need this before we have UI.

We use prod-user-pool Cognito pool here.